refactor return type - change interest rate percent methods to return double

+++ src/main/java/adapter/console/reader/ConsoleInvestmentReaderDelegator.java
@@ -55,7 +55,7 @@ public class ConsoleInvestmentReaderDelegator implements InvestmentReaderDelegat
 	}
 
 	@Override
-	public int readInterestRatePercent(BufferedReader reader) throws IOException {
+	public double readInterestRatePercent(BufferedReader reader) throws IOException {
 		return interestRatePercentReader.read(reader);
 	}
 }
