+++ src/main/java/adapter/console/reader/InterestRatePercentReader.java
@@ -4,5 +4,5 @@ import java.io.BufferedReader;
 import java.io.IOException;
 
 public interface InterestRatePercentReader {
-	int read(BufferedReader reader) throws IOException;
+	double read(BufferedReader reader) throws IOException;
 }
